refactor(handleUri, ClineProvider, zgsmLoginManager): comment out unused Zgsm-related code
